requirements.txt
streamlit==1.29.0
torch==2.2.1
torchvision==0.17.1
Pillow==10.1.0
ultralytics==8.1.0